+++ src/development/vortex/development/networks/modules/postprocess/yolov3.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 from typing import Union, Tuple, List
 
-from .base_postprocess import BasicNMSPostProcess
+from .base_postprocess import BatchedNMSPostProcess
 
 
 def yolo2xywh(bboxes: torch.Tensor):
@@ -17,12 +17,13 @@ def yolo2xywh(bboxes: torch.Tensor):
 
 
 class YoloV3Decoder(nn.Module):
-    def __init__(self, threshold: bool = True):
+    def __init__(self, img_size, threshold: bool = True):
         # note : some backend do not have onnx `NonZero` op,
         # which is necessary for thresholding,
         # `threshold` param allows to turn it off
         super(YoloV3Decoder, self).__init__()
         self.threshold = threshold
+        self.img_size = img_size
 
     def forward(self, input: torch.Tensor, score_threshold: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
         """
@@ -41,7 +42,7 @@ class YoloV3Decoder(nn.Module):
         if self.threshold:
             indices = (predictions[..., 4] > score_threshold)
             indices = indices.squeeze(0)
-            indices = torch.nonzero(indices)
+            indices = torch.nonzero(indices, as_tuple=False)
             predictions = predictions.index_select(1, indices.squeeze(1))
             class_conf = class_conf.index_select(1, indices.squeeze(1))
             class_pred = class_pred.index_select(1, indices.squeeze(1))
@@ -58,7 +59,7 @@ class YoloV3Decoder(nn.Module):
             # predictions = predictions.index_select(1, indices.squeeze(1))
             # class_conf = class_conf.index_select(1, indices.squeeze(1))
             # class_pred = class_pred.index_select(1, indices.squeeze(1))
-        bboxes = predictions[..., :4]
+        bboxes = predictions[..., :4] / self.img_size
         objectness = predictions[..., 4]
         scores = objectness * class_conf.squeeze(2)
         bboxes = yolo2xywh(bboxes)
@@ -66,27 +67,18 @@ class YoloV3Decoder(nn.Module):
         return bboxes.squeeze(0), scores.squeeze(0), class_pred.squeeze(0).squeeze(1), detections
 
 
-class YoloV3PostProcess(BasicNMSPostProcess):
+class YoloV3PostProcess(BatchedNMSPostProcess):
     """ Post-Process for yolo, comply with basic detector post process
     """
 
-    def __init__(self, threshold: bool = True, *args, **kwargs):
+    def __init__(self, img_size, threshold: bool = True, *args, **kwargs):
         super(YoloV3PostProcess, self).__init__(
-            decoder=YoloV3Decoder(threshold=threshold),
+            decoder=YoloV3Decoder(img_size=img_size, threshold=threshold),
             *args, **kwargs
         )
 
 
 def get_postprocess(*args, **kwargs):
     """ return postprocess method for yolo
-    :type *args:
-    :param *args: positional arguments to be forwarded to yolo postprocess
-
-    :type **kwargs:
-    :param **kwargs: keyword arguments to be forwarded to yolo postprocess
-
-    :raises:
-
-    :rtype: YoloPostProcess
     """
     return YoloV3PostProcess(*args, **kwargs)