+++ CHANGELOG.md
@@ -22,7 +22,10 @@ using `'param_groups'` key in model components.
 - validation loss calculation to experiment logger
 - support for changing batch norm layer in all backbone with `norm_layer` argument.
 - new backbone models: CSPNet, ResNest, RexNet, RegNet, TResNet. see more details in [spreadsheet](https://docs.google.com/spreadsheets/d/1M18Bm08P983_-5diHXAmlUmHusjpHdYMmcN0FmrNS74/edit?usp=sharing) or [csv summary](backbones_summary.txt)
-
+- darknet weight converter in [scripts/convert_darknet_weight.py](scripts/convert_darknet_weight.py)
+- yolov3 [converted darknet weight](experiments/outputs/yolov3_darknet_608) and [config file](experiments/configs/yolov3_darknet_608.yml)
+- get model path from `config.checkpoint` if found and valid for `predict` and `export`
+- yolov3 support for changing backbone stage to output with `backbone_stages` argument.
 
 ### Changed
 - model checkpoint not save on hyperparameter optimization
@@ -43,6 +46,7 @@ using `'param_groups'` key in model components.
 - Fix error when `save_best_metrics` not present in experiment file
 - Fix ONNX graph ops to convert `iou_threshold` as input
 - lr scheduler bug when arguments changed and resumed
+- yolov3 model definition
 
 
 ## 0.2.0